refactor(trusty-search): split core/chunker and commands/start into focused submodules (closes #1177, closes #1178)
Part of the line-cap sweep (#607). Both files exceeded the 500-SLOC
production cap; pure mechanical split with zero behavior change.

core/chunker/mod.rs (was 1 534 SLOC → now ~10 SLOC facade):
  - parsers.rs  — per-thread tree-sitter parser cache, ParserKind, parse_with_cached
  - types.rs    — ChunkType, RawChunk, chunk_text
  - document.rs — chunk_document + per-format chunkers (md/yaml/toml/json/txt/xml)
  - ast.rs      — language_for, chunk_ast (public entry point)
  - tests.rs    — all 40 tests (test file, 1 500-SLOC cap)

commands/start.rs (was 1 085 SLOC → deleted, replaced by start/):
  - start/restore.rs  — restore_indexes async warm-boot path
  - start/embedder.rs — build_embedder + adapter types (UDS, lazy, remote)
  - start/daemon.rs   — handle_start boot sequence
  - start/mod.rs      — thin re-export facade
  - start/tests.rs    — all 14 warm-boot + embedder unit tests (test file)

Verification: cargo test -p trusty-search (1 121 tests, 0 failures),
clippy --all-targets -D warnings (clean), fmt --check (clean),
check_line_cap.sh (72 allowlisted, 0 violations, start.rs entry removed).
